Add files via upload
These file are not fully functional. Need to be tested with the database. As for the Function file, it would be the function to calculate the distance between an apartment and the university. However, I could not figure out how to pass properly the parameters that we are getting from the db. we can work on it together